just the next stage in the threefication
