+++ project/plugins.sbt
@@ -1,0 +1,1 @@
+//addSbtPlugin("ch.epfl.scala" % "sbt-scala3-migrate" % "0.4.4")
