+++ scala3/build.sbt
@@ -16,3 +16,32 @@ libraryDependencies += ("org.scala-lang.modules" %% "scala-parser-combinators" %
 libraryDependencies += ("org.specs2" %% "specs2-scalacheck" % "4.12.0" % Test).cross(CrossVersion.for3Use2_13) withSources()
 
 scalacOptions ++= Seq("-deprecation", "-feature")
+
+
+//scalacOptions ++= Seq(
+//  "-deprecation",
+//  "UTF-8",
+//  "-feature")
+
+//scalacOptions ++= {
+//   Seq(
+//      "-encoding",
+//      "UTF-8",
+//      "-feature",
+//      "-language:implicitConversions",
+//      // disabled during the migration
+//      // "-Xfatal-warnings"
+//   ) ++
+//     (CrossVersion.partialVersion(scalaVersion.value) match {
+//        case Some((3, _)) => Seq(
+//           "-unchecked",
+//           "-source:3.0-migration"
+//        )
+//        case _ => Seq(
+//           "-deprecation",
+//           "-Xfatal-warnings",
+//           "-Wunused:imports,privates,locals",
+//           "-Wvalue-discard"
+//        )
+//     })
+//}
